+++ src/main/java/xyz/twooter/configuration/redis/RedisConfiguration.java
@@ -1,4 +1,4 @@
-package xyz.twooter.configuration;
+package xyz.twooter.configuration.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
